Allow for other actions besides routing
Tabs can now be configured to respond and act in other ways than
routing.
The tabs configuration has been moved to the view,
since it mostly deals with presentation matters.
The controller part of the tabs deals with responding to changes in
tabs, either originating internally or from the UI.